WIP on the backend
I studied the Daphne settings (there are still few of them), set up logging.
Digging into Django models, turned core.models into a package.
Finished with the models and with the db router.

Turned core.views into a package.
Added a utility for printing to tty and a script for creating users.
Worked on a test serializer.

Committed on branch 'service'